+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **_Breaking Change_**: remove all `RemoveOnceListener` methods.
   - Use `RemoveListener` instead of `RemoveOnceListener` to remove once listeners.
 - **_Breaking Change_**: remove `CBC.Add` and `UniStartBehaviour.Add`.
+- **_Breaking Change_**: `Entry.GetContext` will search parent first instead of the root object.
 - Feat: add interface `IReadonlyIoC`.
 - Fix: `AdvancedEvent` will call actions and once actions by the order they are added.
   - Previously all once actions will be called after repeated actions.